feat: Enhance product management by integrating image handling and updating API responses

+++ src/app/api/products/[id]/route.ts
@@ -1,5 +1,7 @@
 import { PrismaClient } from '@prisma/client'
-import { NextResponse } from 'next/server'
+import { type NextRequest, NextResponse } from 'next/server'
+
+import { cloudinary } from '@/configs/cloudinary'
 
 const prisma = new PrismaClient()
 
@@ -34,22 +36,54 @@ export async function PUT(
 }
 
 export async function DELETE(
-  request: Request,
+  req: NextRequest,
   { params }: { params: { id: string } }
 ) {
-  const { id } = params
-
   try {
+    const { id } = params
+    console.log(params, 'params')
+    // console.log(req, 'request')
+
+    if (!id) {
+      return NextResponse.json(
+        { success: false, error: 'Product ID is required' },
+        { status: 400 }
+      )
+    }
+
+    // Fetch the product to get its imgId
+    const product = await prisma.product.findUnique({
+      where: { id: Number(id) },
+      include: { image: true },
+    })
+    console.log(product, 'asd')
+
+    if (!product) {
+      return NextResponse.json(
+        { success: false, error: 'Product not found' },
+        { status: 404 }
+      )
+    }
+
+    // Delete the product image from Cloudinary
+    if (product.image?.publicId) {
+      await cloudinary.uploader.destroy(product.image?.publicId)
+      console.log('Image deleted from Cloudinary')
+    } else {
+      console.log('No image to delete', product)
+    }
+
+    // Delete the product from the database
     await prisma.product.delete({
       where: { id: Number(id) },
     })
 
-    return NextResponse.json(
-      { success: true, message: 'Product deleted successfully' },
-      { status: 200 }
-    )
+    return NextResponse.json({
+      success: true,
+      message: 'Product and image deleted successfully',
+    })
   } catch (error) {
-    console.error(error)
+    console.error('Error deleting product:', error)
     return NextResponse.json(
       { success: false, error: 'Failed to delete product' },
       { status: 500 }
